ZAPM Practice mode, and config for testing on debian
Minor improvements to ZAPM practice mode which was added to the
previous (initial) commit, but I forgot to document.

Tournament config file identical to what I am using to test in a
debian container, except for "status" variable which has been set
to "down" for safety.  Please modify according to your environment.