TTYREC recording.
Changed the recording method to ttyrec. This requires ttyrec to be
installed on the server, and path included in nethack_tournament.conf
Added recording of zapm games in the same way - recordings are now
prefixed with z or n for zapm or nethack.  No functionality yet exists
for publishing recordings - I'm open to suggestions.